Remove media capabilities if a server tries to join a conference without putting down an SDP. This scenario has been observed when a corrupted database associated a conference information to a chatroom. At startup, in fact, the core would add media capabilities to the chatroom and if the client is removed from and then added back in, the chatroom zould stay in the CreationPending state because the call_received method would treat the chatroom as such (the received INVITE has no SDP) but method ClientConference::onFocusCallStateChanged would deal with it as it was a conference since it supports media according to its parameters. To overcome this inconsistency, conference parameters are corrected in the ClientConference::onFocusCallStateChanged method. Add chatroom callbacks to conference with chat capabilities when start up a core for testing

+++ liblinphone/src/db/main-db.cpp
@@ -7985,18 +7985,20 @@ void MainDb::insertExpiredConference(const std::shared_ptr<Address> &uri) {
 #endif
 }
 
-void MainDb::deleteConferenceInfo(long long dbConferenceId) {
+void MainDb::deleteConferenceInfo(long long dbConferenceId, bool doCleanup) {
 #ifdef HAVE_DB_STORAGE
 	L_D();
-	long long peerId;
 	soci::session *session = d->dbSession.getBackendSession();
-	std::string peerIdQuery =
-	    "SELECT uri_sip_address_id FROM conference_info WHERE (conference_info.id = :conferenceInfoId)";
-	*session << peerIdQuery, soci::use(dbConferenceId), soci::into(peerId);
-	if (session->got_data()) {
-		long long chatRoomId = d->selectChatRoomId(peerId);
-		if (chatRoomId >= 0) {
-			d->deleteChatRoom(chatRoomId);
+	if (doCleanup) {
+		long long peerId;
+		std::string peerIdQuery =
+		    "SELECT uri_sip_address_id FROM conference_info WHERE (conference_info.id = :conferenceInfoId)";
+		*session << peerIdQuery, soci::use(dbConferenceId), soci::into(peerId);
+		if (session->got_data()) {
+			long long chatRoomId = d->selectChatRoomId(peerId);
+			if (chatRoomId >= 0) {
+				d->deleteChatRoom(chatRoomId);
+			}
 		}
 	}
 
@@ -8005,7 +8007,7 @@ void MainDb::deleteConferenceInfo(long long dbConferenceId) {
 #endif
 }
 
-void MainDb::deleteConferenceInfo(const Address address) {
+void MainDb::deleteConferenceInfo(const Address address, bool doCleanup) {
 #ifdef HAVE_DB_STORAGE
 	if (isInitialized() && address.isValid()) {
 		L_DB_TRANSACTION {
@@ -8013,19 +8015,19 @@ void MainDb::deleteConferenceInfo(const Address address) {
 			lInfo() << "Deleting conference information linked to conference " << address;
 			const long long &uriSipAddressId = d->selectSipAddressId(address, false);
 			const long long &dbConferenceId = d->selectConferenceInfoId(uriSipAddressId);
-			deleteConferenceInfo(dbConferenceId);
+			deleteConferenceInfo(dbConferenceId, doCleanup);
 			tr.commit();
 		};
 	}
 #endif
 }
 
-void MainDb::deleteConferenceInfo(const std::shared_ptr<Address> &address) {
-	deleteConferenceInfo(processConferenceAddress(address));
+void MainDb::deleteConferenceInfo(const std::shared_ptr<Address> &address, bool doCleanup) {
+	deleteConferenceInfo(processConferenceAddress(address), doCleanup);
 }
 
-void MainDb::deleteConferenceInfo(const std::shared_ptr<ConferenceInfo> &conferenceInfo) {
-	deleteConferenceInfo(conferenceInfo->getUri());
+void MainDb::deleteConferenceInfo(const std::shared_ptr<ConferenceInfo> &conferenceInfo, bool doCleanup) {
+	deleteConferenceInfo(conferenceInfo->getUri(), doCleanup);
 }
 
 // -----------------------------------------------------------------------------
